Cleanup: tests, fix ToJSON supeclass.
